Linux (#107)
* Linux 1

* Linux 2

* Linux 3

* Linux 4

* Linux 5

* Linux 6

* Linux 7

* Linux 8

* Linux 9

* Linux 10

* Linux 11

* Linux 12

* Linux 13

* Linux 14

* Linux 15

* Linux 16

* Linux 17

* Linux 18

* Linux 19

* Linux 20

* Linux 21

* Linux 22

* Linux 23

* Linux 24

* Linux 25

* Linux 26

* Linux 27

* Linux 27

* Linux 29

* Linux 30

* Linux 31

* Linux 32

* Linux 33

* Linux 34

* Linux 35

* Linux 36

+++ ltlib/src/io/buffer.h
@@ -1,21 +1,21 @@
 /*
  * BSD 3-Clause License
- * 
+ *
  * Copyright (c) 2023 Zhennan Tu <zhennan.tu@gmail.com>
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice, this
  *    list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- *  
+ *
  * 3. Neither the name of the copyright holder nor the names of its
  *    contributors may be used to endorse or promote products derived from
  *    this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -31,31 +31,23 @@
 #pragma once
 #include <cstdint>
 
-namespace ltlib
-{
+namespace ltlib {
 
-struct Buffer
-{
+struct Buffer {
     explicit Buffer(uint32_t l)
-        : Buffer { new char[l], l }
-    {
-    }
+        : Buffer{new char[l], l} {}
 #ifdef LT_WINDOWS
     Buffer(char* b, uint32_t l)
-        : len { l }
-        , base { b }
-    {
-    }
+        : len{l}
+        , base{b} {}
     uint32_t len;
     char* base;
 #else
     Buffer(char* b, uint32_t l)
-        : base { b }
-        , len { l }
-    {
-    }
+        : base{b}
+        , len{l} {}
     char* base;
-    uint32_t len;
+    size_t len;
 
 #endif
 };
